requirements.txt
flask
pandas
Flask-Caching
spacy
imapclient==2.3.1
email-validator==2.1.0.post1
pandas==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
spacy==3.6.0
word2number==1.1
scikit-learn==1.3.0
numpy==1.25.2
regex==2023.8.8
Pillow==10.0.0
